refactor: extract getInterpolation in its own file
